refactoring - change name of package from models.dao to persistance
