+++ e-deansoffice/src/main/java/pl/edu/agh/edeansoffice/persistance/FieldsOfStudyDao.java
@@ -1,0 +1,20 @@
+package pl.edu.agh.edeansoffice.persistance;
+
+import java.util.List;
+
+import javax.transaction.Transactional;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import pl.edu.agh.edeansoffice.models.FieldsOfStudy;
+
+@Transactional
+@Repository
+public interface FieldsOfStudyDao extends CrudRepository<FieldsOfStudy, Integer> {
+	
+	public List<FieldsOfStudy> findAll();
+	
+	public FieldsOfStudy findById(Integer id);
+
+}
