+++ e-deansoffice/src/main/java/pl/edu/agh/edeansoffice/persistance/StudentsDao.java
@@ -1,0 +1,23 @@
+package pl.edu.agh.edeansoffice.persistance;
+
+import java.util.List;
+
+import javax.transaction.Transactional;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import pl.edu.agh.edeansoffice.models.Students;
+
+@Transactional
+@Repository
+public interface StudentsDao extends CrudRepository<Students, Integer> {
+	
+	public List<Students> findAll();
+	
+	public Students findById(Integer id);
+	
+	public Students findByStudentsBookNumber(String studentsBookNumber);
+	
+	public Students findByPassword(String password);
+}
