+++ e-deansoffice/src/main/java/pl/edu/agh/edeansoffice/persistance/TeachersDao.java
@@ -1,0 +1,19 @@
+package pl.edu.agh.edeansoffice.persistance;
+
+import java.util.List;
+
+import javax.transaction.Transactional;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import pl.edu.agh.edeansoffice.models.Teachers;
+
+@Transactional
+@Repository
+public interface TeachersDao extends CrudRepository<Teachers, Integer> {
+	
+	public List<Teachers> findAll();
+	
+	public Teachers findById(Integer id);
+}
